Update for new Slack apps with granular scopes
Use the new manifest to configure the app and specify the scopes
required.

Use the built-in install app button to obtain access token. This removes
the need for the install and authorized endpoints, along with the
DynamoDB table to store access tokens. It also simplies the env vars
from client id and secret to just the access token itself. This does
prevent the bot being used across teams, but this is not currently the
preferred deployment scenario.

The event for bot messages appears to have changed slightly with new
Slack apps, presumably driven by the access token endpoint version. Now
identify bot messages using bot_id rather than by subtype.